+++ arithmetization/src/test/examples/zig/build.zig
@@ -22,8 +22,8 @@ pub fn build(b: *std.Build) void {
         .preferred_optimize_mode = .ReleaseSmall,
     });
 
-    const name = b.option([]const u8, "name", "Name of the program (source: src/<name>.zig, binary: <name>)") orelse @panic("'-Dname=<name>' is required");
-    const strip = b.option(bool, "strip", "Whether to strip the binary (default: false)") orelse false;
+    const path = b.option([]const u8, "path", "Source path under src/, without .zig")
+    orelse @panic("'-Dpath=<path>' is required");
 
     // e.g. path = "src_optional_subfolder/your_test"
     const source = b.fmt("src/{s}.zig", .{path});
@@ -32,17 +32,24 @@ pub fn build(b: *std.Build) void {
     const exe_name = std.fs.path.stem(std.fs.path.basename(path));
 
     const exe = b.addExecutable(.{
-        .name = name,
+        .name = exe_name,
         .root_module = b.createModule(.{
             .root_source_file = b.path(source),
             .target = target,
             .optimize = optimize,
-            .strip = strip, // Removes debug symbols and other metadata
         }),
     });
 
-    // Point to assembly overwriting default SP
+    const wrappers = b.createModule(.{
+        .root_source_file = b.path("../../../wrappers/root.zig"),
+        .target = target,
+        .optimize = optimize,
+    });
+    exe.root_module.addImport("wrappers", wrappers);
+
+    // Point to assembly overwriting default SP with the one defined in the linker script
     exe.root_module.addAssemblyFile(b.path("src/start.s"));
+    exe.setLinkerScript(b.path("../linker_script.ld"));
 
     // Remove unused code sections
     exe.link_gc_sections = true;